Compile error on multiple `#[divan::bench]` use
This statically prevents the attribute from being used multiple times on
the same attribute.

Coincidentally, this change also causes the `weird_usage::unique_id`
test to fail for `lifetime` and  `embedded::inner::inner`. It appears
that type ID uniqueness is busted for functions with lifetimes. This is
a known issue: https://github.com/rust-lang/rust/issues/84366.