Support generic types in `#[divan::bench]`
Given `#[divan::bench(types = [A, B, C])]`, each concrete type will get
its own benchmark. Currently, only functions with a single generic type
are supported.

The attribute generates a `GroupEntry` for the function with a list of
`GenericBenchEntry` for each concrete type. As a result, they are
discovered via `GROUP_ENTRIES` instead of `BENCH_ENTRIES`.

`GenericBenchEntry` contains a function for getting the `TypeId`. It is
currently unused, but may be used in the future for querying benchmark
functions for a specific type.