Move `#[ignore]` to `BenchOptions`
This allows for instead doing `#[divan::bench(ignore)]` or
`#[divan::bench_group(ignore)]`. Because we use field update syntax for
`BenchOptions`, the compiler automatically prevents the option from
being set in both places.

This also removes the need to handle `parent_ignored` when traversing
the `EntryTree` and we can instead rely on `BenchOptions::overwrite`.