Format output as table right of entry tree
The table contains the following stats as latency and optionally with
counter throughput below:
- Fastest duration
- Slowest duration
- Median duration
- Mean duration

This removes the option for terse formatting since that was half-baked
and relied on `Stats` debug formatting. Terse formatting will be added
later once it is more well thought out.

Changed `FineDuration` formatting to use a space between the value and
unit suffix.