Reintroduce counter to search example benchmarks
Change 1713c6b (#24) removed the use of the `ItemsCount` counter because
the resulting numbers didn't make much sense. So it was unreasonable to
use the haystack size.

Every benchmark iteration searches for a single element, so it appears
that the correct count to use is 1. The new throughput numbers now make
more sense with this change.