Reduce `AllocProfiler` footprint on x86_64 macOS
Adds a fast path for accessing the local `ThreadAllocInfo` instance via
a static key instead of one allocated from `pthread_key_create`. This
removes a global memory load that may not be cached. I chose key 11
because reserves it for ABI compatibility with Windows.

It can be opted out of by enabling the `dyn_thread_local` crate feature.
This allows for other libraries to use this key statically.